api/requirements.txt
fastapi==0.121.3
uvicorn==0.38.0
langgraph==1.0.3
pydantic==2.12.4
openai-agents==0.6.1
python-dotenv==1.2.1
twilio==9.8.7
